requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.7.0
rich>=13.7.0